@@ -1,30 +1,34 @@
 {
     "name": "wikia/prometheus_client_php",
+    "description": "Prometheus instrumentation library for PHP applications.",
+    "type": "library",
+    "license": "Apache-2.0",
     "authors": [
         {
-            "name": "Joscha",
-            "email": "joscha@schnipseljagd.org"
-        },
-        {
-            "name": "Jan Brauer",
-            "email": "jan.brauer@jimdo.com"
+            "name": "Daniel Noel-Davies",
+            "email": "Daniel.Noel-Davies@endclothing.com"
         }
     ],
+    "replace": {
+        "jimdo/prometheus_client_php": "*"
+    },
     "require": {
-        "php": ">=7.0",
-        "guzzlehttp/guzzle": "^6.2",
+        "php": "^7.2",
+        "ext-json": "*",
+        "guzzlehttp/guzzle": "^6.3",
         "symfony/polyfill-apcu": "^1.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.1.0"
+        "phpunit/phpunit": "^8.4",
+        "squizlabs/php_codesniffer": "^3.5"
     },
     "suggest": {
         "ext-redis": "Required if using Redis.",
         "ext-apc": "Required if using APCu."
     },
     "autoload": {
-        "psr-0": {
-            "Prometheus\\": "src/"
+        "psr-4": {
+            "Prometheus\\": "src/Prometheus/"
         }
     },
     "autoload-dev": {
@@ -32,5 +36,13 @@
             "Test\\Prometheus\\": "tests/"
         }
     },
-    "license": "Apache-2.0"
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0-dev"
+        }
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "prefer-stable": true
 }
