@@ -9,7 +9,7 @@ use Prometheus\Storage\Redis;
 $adapter = $_GET['adapter'];
 
 if ($adapter === 'redis') {
-    Redis::setDefaultOptions(array('host' => isset($_SERVER['REDIS_HOST']) ? $_SERVER['REDIS_HOST'] : '127.0.0.1'));
+    Redis::setDefaultOptions(['host' => $_SERVER['REDIS_HOST'] ?? '127.0.0.1']);
     $adapter = new Prometheus\Storage\Redis();
 } elseif ($adapter === 'apc') {
     $adapter = new Prometheus\Storage\APC();
