@@ -6,12 +6,12 @@ use Prometheus\CollectorRegistry;
 use Prometheus\Storage\Redis;
 
 
-error_log('c='. $_GET['c']);
+error_log('c=' . $_GET['c']);
 
 $adapter = $_GET['adapter'];
 
 if ($adapter === 'redis') {
-    Redis::setDefaultOptions(array('host' => isset($_SERVER['REDIS_HOST']) ? $_SERVER['REDIS_HOST'] : '127.0.0.1'));
+    Redis::setDefaultOptions(['host' => $_SERVER['REDIS_HOST'] ?? '127.0.0.1']);
     $adapter = new Prometheus\Storage\Redis();
 } elseif ($adapter === 'apc') {
     $adapter = new Prometheus\Storage\APC();
