@@ -1,8 +1,8 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Prometheus;
-
 
 use Prometheus\Exception\MetricNotFoundException;
 use Prometheus\Exception\MetricsRegistrationException;
@@ -20,19 +20,26 @@ class CollectorRegistry
      * @var Adapter
      */
     private $storageAdapter;
+
     /**
      * @var Gauge[]
      */
-    private $gauges = array();
+    private $gauges = [];
+
     /**
      * @var Counter[]
      */
-    private $counters = array();
+    private $counters = [];
+
     /**
      * @var Histogram[]
      */
-    private $histograms = array();
+    private $histograms = [];
 
+    /**
+     * CollectorRegistry constructor.
+     * @param Adapter $redisAdapter
+     */
     public function __construct(Adapter $redisAdapter)
     {
         $this->storageAdapter = $redisAdapter;
@@ -41,7 +48,7 @@ class CollectorRegistry
     /**
      * @return CollectorRegistry
      */
-    public static function getDefault()
+    public static function getDefault(): CollectorRegistry
     {
         if (!self::$defaultRegistry) {
             return self::$defaultRegistry = new static(new Redis());
@@ -52,7 +59,7 @@ class CollectorRegistry
     /**
      * @return MetricFamilySamples[]
      */
-    public function getMetricFamilySamples()
+    public function getMetricFamilySamples(): array
     {
         return $this->storageAdapter->collect();
     }
@@ -65,7 +72,7 @@ class CollectorRegistry
      * @return Gauge
      * @throws MetricsRegistrationException
      */
-    public function registerGauge($namespace, $name, $help, $labels = array())
+    public function registerGauge($namespace, $name, $help, $labels = []): Gauge
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (isset($this->gauges[$metricIdentifier])) {
@@ -87,7 +94,7 @@ class CollectorRegistry
      * @return Gauge
      * @throws MetricNotFoundException
      */
-    public function getGauge($namespace, $name)
+    public function getGauge($namespace, $name): Gauge
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (!isset($this->gauges[$metricIdentifier])) {
@@ -102,8 +109,9 @@ class CollectorRegistry
      * @param string $help e.g. The duration something took in seconds.
      * @param array $labels e.g. ['controller', 'action']
      * @return Gauge
+     * @throws MetricsRegistrationException
      */
-    public function getOrRegisterGauge($namespace, $name, $help, $labels = array())
+    public function getOrRegisterGauge($namespace, $name, $help, $labels = []): Gauge
     {
         try {
             $gauge = $this->getGauge($namespace, $name);
@@ -121,7 +129,7 @@ class CollectorRegistry
      * @return Counter
      * @throws MetricsRegistrationException
      */
-    public function registerCounter($namespace, $name, $help, $labels = array())
+    public function registerCounter($namespace, $name, $help, $labels = []): Counter
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (isset($this->counters[$metricIdentifier])) {
@@ -143,7 +151,7 @@ class CollectorRegistry
      * @return Counter
      * @throws MetricNotFoundException
      */
-    public function getCounter($namespace, $name)
+    public function getCounter($namespace, $name): Counter
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (!isset($this->counters[$metricIdentifier])) {
@@ -158,8 +166,9 @@ class CollectorRegistry
      * @param string $help e.g. The number of requests made.
      * @param array $labels e.g. ['controller', 'action']
      * @return Counter
+     * @throws MetricsRegistrationException
      */
-    public function getOrRegisterCounter($namespace, $name, $help, $labels = array())
+    public function getOrRegisterCounter($namespace, $name, $help, $labels = []): Counter
     {
         try {
             $counter = $this->getCounter($namespace, $name);
@@ -178,7 +187,7 @@ class CollectorRegistry
      * @return Histogram
      * @throws MetricsRegistrationException
      */
-    public function registerHistogram($namespace, $name, $help, $labels = array(), $buckets = null)
+    public function registerHistogram($namespace, $name, $help, $labels = [], $buckets = null): Histogram
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (isset($this->histograms[$metricIdentifier])) {
@@ -201,7 +210,7 @@ class CollectorRegistry
      * @return Histogram
      * @throws MetricNotFoundException
      */
-    public function getHistogram($namespace, $name)
+    public function getHistogram($namespace, $name): Histogram
     {
         $metricIdentifier = self::metricIdentifier($namespace, $name);
         if (!isset($this->histograms[$metricIdentifier])) {
@@ -217,8 +226,9 @@ class CollectorRegistry
      * @param array $labels e.g. ['controller', 'action']
      * @param array $buckets e.g. [100, 200, 300]
      * @return Histogram
+     * @throws MetricsRegistrationException
      */
-    public function getOrRegisterHistogram($namespace, $name, $help, $labels = array(), $buckets = null)
+    public function getOrRegisterHistogram($namespace, $name, $help, $labels = [], $buckets = null): Histogram
     {
         try {
             $histogram = $this->getHistogram($namespace, $name);
@@ -228,7 +238,12 @@ class CollectorRegistry
         return $histogram;
     }
 
-    private static function metricIdentifier($namespace, $name)
+    /**
+     * @param $namespace
+     * @param $name
+     * @return string
+     */
+    private static function metricIdentifier($namespace, $name): string
     {
         return $namespace . ":" . $name;
     }
