@@ -1,14 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Prometheus;
 
-
+use InvalidArgumentException;
 use Prometheus\Storage\Adapter;
 
 class Histogram extends Collector
 {
     const TYPE = 'histogram';
 
+    /**
+     * @var array|null
+     */
     private $buckets;
 
     /**
@@ -19,7 +24,7 @@ class Histogram extends Collector
      * @param array $labels
      * @param array $buckets
      */
-    public function __construct(Adapter $adapter, $namespace, $name, $help, $labels = array(), $buckets = null)
+    public function __construct(Adapter $adapter, $namespace, $name, $help, $labels = [], $buckets = null)
     {
         parent::__construct($adapter, $namespace, $name, $help, $labels);
 
@@ -28,21 +33,19 @@ class Histogram extends Collector
         }
 
         if (0 == count($buckets)) {
-            throw new \InvalidArgumentException("Histogram must have at least one bucket.");
+            throw new InvalidArgumentException("Histogram must have at least one bucket.");
         }
 
         for ($i = 0; $i < count($buckets) - 1; $i++) {
             if ($buckets[$i] >= $buckets[$i + 1]) {
-                throw new \InvalidArgumentException(
+                throw new InvalidArgumentException(
                     "Histogram buckets must be in increasing order: " .
                     $buckets[$i] . " >= " . $buckets[$i + 1]
                 );
             }
         }
-        foreach ($labels as $label) {
-            if ($label === 'le') {
-                throw new \InvalidArgumentException("Histogram cannot have a label named 'le'.");
-            }
+        if (in_array('le', $labels, true)) {
+            throw new \InvalidArgumentException("Histogram cannot have a label named 'le'.");
         }
         $this->buckets = $buckets;
     }
@@ -51,23 +54,53 @@ class Histogram extends Collector
      * List of default buckets suitable for typical web application latency metrics
      * @return array
      */
-    public static function getDefaultBuckets()
+    public static function getDefaultBuckets(): array
     {
-        return array(
-            0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0
-        );
+        return [
+            0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0,
+        ];
+    }
+
+    /**
+     * @param float $start
+     * @param float $growthFactor
+     * @param int $numberOfBuckets
+     * @return array
+     */
+    public static function exponentialBuckets(float $start, float $growthFactor, int $numberOfBuckets): array
+    {
+        if ($numberOfBuckets < 1) {
+            throw new InvalidArgumentException('Number of buckets must be a positive integer');
+        }
+
+        if ($start <= 0) {
+            throw new InvalidArgumentException('The starting position of a set of buckets must be a positive integer');
+        }
+
+        if ($growthFactor <= 1) {
+            throw new InvalidArgumentException('The growth factor must greater than 1');
+        }
+
+        $buckets = [];
+
+        for ($i = 0; $i < $numberOfBuckets; $i++) {
+            $buckets[$i] = $start;
+            $start *= $growthFactor;
+        }
+
+        return $buckets;
     }
 
     /**
      * @param double $value e.g. 123
      * @param array $labels e.g. ['status', 'opcode']
      */
-    public function observe($value, $labels = array())
+    public function observe(float $value, array $labels = []): void
     {
         $this->assertLabelsAreDefinedCorrectly($labels);
 
         $this->storageAdapter->updateHistogram(
-            array(
+            [
                 'value' => $value,
                 'name' => $this->getName(),
                 'help' => $this->getHelp(),
@@ -75,14 +108,14 @@ class Histogram extends Collector
                 'labelNames' => $this->getLabelNames(),
                 'labelValues' => $labels,
                 'buckets' => $this->buckets,
-            )
+            ]
         );
     }
 
     /**
      * @return string
      */
-    public function getType()
+    public function getType(): string
     {
         return self::TYPE;
     }
