@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Prometheus\Storage;
 
-
+use InvalidArgumentException;
 use Prometheus\Counter;
 use Prometheus\Exception\StorageException;
 use Prometheus\Gauge;
@@ -13,53 +15,81 @@ class Redis implements Adapter
 {
     const PROMETHEUS_METRIC_KEYS_SUFFIX = '_METRIC_KEYS';
 
-    private static $defaultOptions = array();
+    /**
+     * @var array
+     */
+    private static $defaultOptions = [
+        'host' => '127.0.0.1',
+        'port' => 6379,
+        'timeout' => 0.1,
+        'read_timeout' => '10',
+        'persistent_connections' => false,
+        'password' => null,
+    ];
+
+    /**
+     * @var string
+     */
     private static $prefix = 'PROMETHEUS_';
 
-    private $options;
+    /**
+     * @var array
+     */
+    private $options = [];
+
+    /**
+     * @var \Redis
+     */
     private $redis;
 
-    public function __construct(array $options = array())
-    {
-        // with php 5.3 we cannot initialize the options directly on the field definition
-        // so we initialize them here for now
-        if (!isset(self::$defaultOptions['host'])) {
-            self::$defaultOptions['host'] = '127.0.0.1';
-        }
-        if (!isset(self::$defaultOptions['port'])) {
-            self::$defaultOptions['port'] = 6379;
-        }
-        if (!isset(self::$defaultOptions['timeout'])) {
-            self::$defaultOptions['timeout'] = 0.1; // in seconds
-        }
-        if (!isset(self::$defaultOptions['read_timeout'])) {
-            self::$defaultOptions['read_timeout'] = 10; // in seconds
-        }
-        if (!isset(self::$defaultOptions['persistent_connections'])) {
-            self::$defaultOptions['persistent_connections'] = false;
-        }
-        if (!isset(self::$defaultOptions['password'])) {
-            self::$defaultOptions['password'] = null;
-        }
+    /**
+     * @var boolean
+     */
+    private $connectionInitialized = false;
 
+    /**
+     * Redis constructor.
+     * @param array $options
+     */
+    public function __construct(array $options = [])
+    {
         $this->options = array_merge(self::$defaultOptions, $options);
         $this->redis = new \Redis();
+    }
+
+    public static function fromExistingConnection(\Redis $redis): self
+    {
+        if ($redis->isConnected() === false) {
+            throw new StorageException('Connection to Redis server not established');
+        }
+
+        $self = new self();
+        $self->connectionInitialized = true;
+        $self->redis = $redis;
+
+        return $self;
     }
 
     /**
      * @param array $options
      */
-    public static function setDefaultOptions(array $options)
+    public static function setDefaultOptions(array $options): void
     {
         self::$defaultOptions = array_merge(self::$defaultOptions, $options);
     }
 
-    public static function setPrefix($prefix)
+    /**
+     * @param $prefix
+     */
+    public static function setPrefix($prefix): void
     {
         self::$prefix = $prefix;
     }
 
-    public function flushRedis()
+    /**
+     * @throws StorageException
+     */
+    public function flushRedis(): void
     {
         $this->openConnection();
         $this->redis->flushAll();
@@ -69,7 +99,7 @@ class Redis implements Adapter
      * @return MetricFamilySamples[]
      * @throws StorageException
      */
-    public function collect()
+    public function collect(): array
     {
         $this->openConnection();
         $metrics = $this->collectHistograms();
@@ -86,24 +116,53 @@ class Redis implements Adapter
     /**
      * @throws StorageException
      */
-    private function openConnection()
+    private function openConnection(): void
+    {
+        if ($this->connectionInitialized === true) {
+            return;
+        }
+
+        $connectionStatus = $this->connectToServer();
+        if ($connectionStatus === false) {
+            throw new StorageException("Can't connect to Redis server", 0);
+        }
+
+        if ($this->options['password']) {
+            $this->redis->auth($this->options['password']);
+        }
+
+        if (isset($this->options['database'])) {
+            $this->redis->select($this->options['database']);
+        }
+
+        $this->redis->setOption(\Redis::OPT_READ_TIMEOUT, $this->options['read_timeout']);
+    }
+
+    /**
+     * @return bool
+     */
+    private function connectToServer(): bool
     {
         try {
             if ($this->options['persistent_connections']) {
-                @$this->redis->pconnect($this->options['host'], $this->options['port'], $this->options['timeout']);
-            } else {
-                @$this->redis->connect($this->options['host'], $this->options['port'], $this->options['timeout']);
+                return $this->redis->pconnect(
+                    $this->options['host'],
+                    $this->options['port'],
+                    $this->options['timeout']
+                );
             }
-            if ($this->options['password']) {
-                $this->redis->auth($this->options['password']);
-            }
-            $this->redis->setOption(\Redis::OPT_READ_TIMEOUT, $this->options['read_timeout']);
+
+            return $this->redis->connect($this->options['host'], $this->options['port'], $this->options['timeout']);
         } catch (\RedisException $e) {
-            throw new StorageException("Can't connect to Redis server", 0, $e);
+            return false;
         }
     }
 
-    public function updateHistogram(array $data)
+    /**
+     * @param array $data
+     * @throws StorageException
+     */
+    public function updateHistogram(array $data): void
     {
         $this->openConnection();
         $bucketToIncrease = '+Inf';
@@ -116,106 +175,120 @@ class Redis implements Adapter
         $metaData = $data;
         unset($metaData['value']);
         unset($metaData['labelValues']);
-        $this->redis->eval(<<<LUA
-local increment = redis.call('hIncrByFloat', KEYS[1], KEYS[2], ARGV[1])
-redis.call('hIncrBy', KEYS[1], KEYS[3], 1)
-if increment == ARGV[1] then
-    redis.call('hSet', KEYS[1], '__meta', ARGV[2])
-    redis.call('sAdd', KEYS[4], KEYS[1])
+
+        $this->redis->eval(
+            <<<LUA
+local increment = redis.call('hIncrByFloat', KEYS[1], ARGV[1], ARGV[3])
+redis.call('hIncrBy', KEYS[1], ARGV[2], 1)
+if increment == ARGV[3] then
+    redis.call('hSet', KEYS[1], '__meta', ARGV[4])
+    redis.call('sAdd', KEYS[2], KEYS[1])
 end
 LUA
             ,
-            array(
+            [
                 $this->toMetricKey($data),
-                json_encode(array('b' => 'sum', 'labelValues' => $data['labelValues'])),
-                json_encode(array('b' => $bucketToIncrease, 'labelValues' => $data['labelValues'])),
                 self::$prefix . Histogram::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX,
+                json_encode(['b' => 'sum', 'labelValues' => $data['labelValues']]),
+                json_encode(['b' => $bucketToIncrease, 'labelValues' => $data['labelValues']]),
                 $data['value'],
                 json_encode($metaData),
-            ),
-            4
+            ],
+            2
         );
     }
 
-    public function updateGauge(array $data)
+    /**
+     * @param array $data
+     * @throws StorageException
+     */
+    public function updateGauge(array $data): void
     {
         $this->openConnection();
         $metaData = $data;
         unset($metaData['value']);
         unset($metaData['labelValues']);
         unset($metaData['command']);
-        $this->redis->eval(<<<LUA
-local result = redis.call(KEYS[2], KEYS[1], KEYS[4], ARGV[1])
+        $this->redis->eval(
+            <<<LUA
+local result = redis.call(ARGV[1], KEYS[1], ARGV[2], ARGV[3])
 
-if KEYS[2] == 'hSet' then
+if ARGV[1] == 'hSet' then
     if result == 1 then
-        redis.call('hSet', KEYS[1], '__meta', ARGV[2])
-        redis.call('sAdd', KEYS[3], KEYS[1])
+        redis.call('hSet', KEYS[1], '__meta', ARGV[4])
+        redis.call('sAdd', KEYS[2], KEYS[1])
     end
 else
-    if result == ARGV[1] then
-        redis.call('hSet', KEYS[1], '__meta', ARGV[2])
-        redis.call('sAdd', KEYS[3], KEYS[1])
+    if result == ARGV[3] then
+        redis.call('hSet', KEYS[1], '__meta', ARGV[4])
+        redis.call('sAdd', KEYS[2], KEYS[1])
     end
 end
 LUA
             ,
-            array(
+            [
                 $this->toMetricKey($data),
-                $this->getRedisCommand($data['command']),
                 self::$prefix . Gauge::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX,
+                $this->getRedisCommand($data['command']),
                 json_encode($data['labelValues']),
                 $data['value'],
                 json_encode($metaData),
-            ),
-            4
+            ],
+            2
         );
     }
 
-    public function updateCounter(array $data)
+    /**
+     * @param array $data
+     * @throws StorageException
+     */
+    public function updateCounter(array $data): void
     {
         $this->openConnection();
         $metaData = $data;
         unset($metaData['value']);
         unset($metaData['labelValues']);
         unset($metaData['command']);
-        $result = $this->redis->eval(<<<LUA
-local result = redis.call(KEYS[2], KEYS[1], KEYS[4], ARGV[1])
-if result == tonumber(ARGV[1]) then
-    redis.call('hMSet', KEYS[1], '__meta', ARGV[2])
-    redis.call('sAdd', KEYS[3], KEYS[1])
+        $this->redis->eval(
+            <<<LUA
+local result = redis.call(ARGV[1], KEYS[1], ARGV[3], ARGV[2])
+if result == tonumber(ARGV[2]) then
+    redis.call('hMSet', KEYS[1], '__meta', ARGV[4])
+    redis.call('sAdd', KEYS[2], KEYS[1])
 end
 return result
 LUA
             ,
-            array(
+            [
                 $this->toMetricKey($data),
-                $this->getRedisCommand($data['command']),
                 self::$prefix . Counter::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX,
-                json_encode($data['labelValues']),
+                $this->getRedisCommand($data['command']),
                 $data['value'],
+                json_encode($data['labelValues']),
                 json_encode($metaData),
-            ),
-            4
+            ],
+            2
         );
-        return $result;
     }
 
-    private function collectHistograms()
+    /**
+     * @return array
+     */
+    private function collectHistograms(): array
     {
         $keys = $this->redis->sMembers(self::$prefix . Histogram::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX);
         sort($keys);
-        $histograms = array();
+        $histograms = [];
         foreach ($keys as $key) {
-            $raw = $this->redis->hGetAll($key);
+            $raw = $this->redis->hGetAll(str_replace($this->redis->_prefix(''), '', $key));
             $histogram = json_decode($raw['__meta'], true);
             unset($raw['__meta']);
-            $histogram['samples'] = array();
+            $histogram['samples'] = [];
 
             // Add the Inf bucket so we can compute it later on
             $histogram['buckets'][] = '+Inf';
 
-            $allLabelValues = array();
+            $allLabelValues = [];
             foreach (array_keys($raw) as $k) {
                 $d = json_decode($k, true);
                 if ($d['b'] == 'sum') {
@@ -235,65 +308,68 @@ LUA
                 // the previous one.
                 $acc = 0;
                 foreach ($histogram['buckets'] as $bucket) {
-                    $bucketKey = json_encode(array('b' => $bucket, 'labelValues' => $labelValues));
+                    $bucketKey = json_encode(['b' => $bucket, 'labelValues' => $labelValues]);
                     if (!isset($raw[$bucketKey])) {
-                        $histogram['samples'][] = array(
+                        $histogram['samples'][] = [
                             'name' => $histogram['name'] . '_bucket',
-                            'labelNames' => array('le'),
-                            'labelValues' => array_merge($labelValues, array($bucket)),
-                            'value' => $acc
-                        );
+                            'labelNames' => ['le'],
+                            'labelValues' => array_merge($labelValues, [$bucket]),
+                            'value' => $acc,
+                        ];
                     } else {
                         $acc += $raw[$bucketKey];
-                        $histogram['samples'][] = array(
+                        $histogram['samples'][] = [
                             'name' => $histogram['name'] . '_bucket',
-                            'labelNames' => array('le'),
-                            'labelValues' => array_merge($labelValues, array($bucket)),
-                            'value' => $acc
-                        );
+                            'labelNames' => ['le'],
+                            'labelValues' => array_merge($labelValues, [$bucket]),
+                            'value' => $acc,
+                        ];
                     }
                 }
 
                 // Add the count
-                $histogram['samples'][] = array(
+                $histogram['samples'][] = [
                     'name' => $histogram['name'] . '_count',
-                    'labelNames' => array(),
+                    'labelNames' => [],
                     'labelValues' => $labelValues,
-                    'value' => $acc
-                );
+                    'value' => $acc,
+                ];
 
                 // Add the sum
-                $histogram['samples'][] = array(
+                $histogram['samples'][] = [
                     'name' => $histogram['name'] . '_sum',
-                    'labelNames' => array(),
+                    'labelNames' => [],
                     'labelValues' => $labelValues,
-                    'value' => $raw[json_encode(array('b' => 'sum', 'labelValues' => $labelValues))]
-                );
+                    'value' => $raw[json_encode(['b' => 'sum', 'labelValues' => $labelValues])],
+                ];
             }
             $histograms[] = $histogram;
         }
         return $histograms;
     }
 
-    private function collectGauges()
+    /**
+     * @return array
+     */
+    private function collectGauges(): array
     {
         $keys = $this->redis->sMembers(self::$prefix . Gauge::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX);
         sort($keys);
-        $gauges = array();
+        $gauges = [];
         foreach ($keys as $key) {
-            $raw = $this->redis->hGetAll($key);
+            $raw = $this->redis->hGetAll(str_replace($this->redis->_prefix(''), '', $key));
             $gauge = json_decode($raw['__meta'], true);
             unset($raw['__meta']);
-            $gauge['samples'] = array();
+            $gauge['samples'] = [];
             foreach ($raw as $k => $value) {
-                $gauge['samples'][] = array(
+                $gauge['samples'][] = [
                     'name' => $gauge['name'],
-                    'labelNames' => array(),
+                    'labelNames' => [],
                     'labelValues' => json_decode($k, true),
-                    'value' => $value
-                );
+                    'value' => $value,
+                ];
             }
-            usort($gauge['samples'], function($a, $b){
+            usort($gauge['samples'], function ($a, $b) {
                 return strcmp(implode("", $a['labelValues']), implode("", $b['labelValues']));
             });
             $gauges[] = $gauge;
@@ -301,25 +377,28 @@ LUA
         return $gauges;
     }
 
-    private function collectCounters()
+    /**
+     * @return array
+     */
+    private function collectCounters(): array
     {
         $keys = $this->redis->sMembers(self::$prefix . Counter::TYPE . self::PROMETHEUS_METRIC_KEYS_SUFFIX);
         sort($keys);
-        $counters = array();
+        $counters = [];
         foreach ($keys as $key) {
-            $raw = $this->redis->hGetAll($key);
+            $raw = $this->redis->hGetAll(str_replace($this->redis->_prefix(''), '', $key));
             $counter = json_decode($raw['__meta'], true);
             unset($raw['__meta']);
-            $counter['samples'] = array();
+            $counter['samples'] = [];
             foreach ($raw as $k => $value) {
-                $counter['samples'][] = array(
+                $counter['samples'][] = [
                     'name' => $counter['name'],
-                    'labelNames' => array(),
+                    'labelNames' => [],
                     'labelValues' => json_decode($k, true),
-                    'value' => $value
-                );
+                    'value' => $value,
+                ];
             }
-            usort($counter['samples'], function($a, $b){
+            usort($counter['samples'], function ($a, $b) {
                 return strcmp(implode("", $a['labelValues']), implode("", $b['labelValues']));
             });
             $counters[] = $counter;
@@ -327,7 +406,11 @@ LUA
         return $counters;
     }
 
-    private function getRedisCommand($cmd)
+    /**
+     * @param int $cmd
+     * @return string
+     */
+    private function getRedisCommand(int $cmd): string
     {
         switch ($cmd) {
             case Adapter::COMMAND_INCREMENT_INTEGER:
@@ -337,7 +420,7 @@ LUA
             case Adapter::COMMAND_SET:
                 return 'hSet';
             default:
-                throw new \InvalidArgumentException("Unknown command");
+                throw new InvalidArgumentException("Unknown command");
         }
     }
 
@@ -345,9 +428,8 @@ LUA
      * @param array $data
      * @return string
      */
-    private function toMetricKey(array $data)
+    private function toMetricKey(array $data): string
     {
-        return implode(':', array(self::$prefix, $data['type'], $data['name']));
+        return implode(':', [self::$prefix, $data['type'], $data['name']]);
     }
-
 }
