@@ -1,11 +1,12 @@
 <?php
+
 namespace Test;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Promise;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class BlackBoxTest extends PHPUnit_Framework_TestCase
+class BlackBoxTest extends TestCase
 {
     /**
      * @var Client
