@@ -1,9 +1,9 @@
 <?php
 
-
 namespace Test\Prometheus;
 
-use PHPUnit_Framework_TestCase;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 use Prometheus\Counter;
 use Prometheus\MetricFamilySamples;
 use Prometheus\Sample;
@@ -12,48 +12,50 @@ use Prometheus\Storage\Adapter;
 /**
  * See https://prometheus.io/docs/instrumenting/exposition_formats/
  */
-abstract class AbstractCounterTest extends PHPUnit_Framework_TestCase
+abstract class AbstractCounterTest extends TestCase
 {
     /**
      * @var Adapter
      */
     public $adapter;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->configureAdapter();
     }
+
+    abstract public function configureAdapter();
 
     /**
      * @test
      */
     public function itShouldIncreaseWithLabels()
     {
-        $gauge = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing', array('foo', 'bar'));
-        $gauge->inc(array('lalal', 'lululu'));
-        $gauge->inc(array('lalal', 'lululu'));
-        $gauge->inc(array('lalal', 'lululu'));
+        $counter = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing', ['foo', 'bar']);
+        $counter->inc(['lalal', 'lululu']);
+        $counter->inc(['lalal', 'lululu']);
+        $counter->inc(['lalal', 'lululu']);
         $this->assertThat(
             $this->adapter->collect(),
             $this->equalTo(
-                array(
+                [
                     new MetricFamilySamples(
-                        array(
+                        [
                             'type' => Counter::TYPE,
                             'help' => 'this is for testing',
                             'name' => 'test_some_metric',
-                            'labelNames' => array('foo', 'bar'),
-                            'samples' => array(
-                                array(
-                                    'labelValues' => array('lalal', 'lululu'),
+                            'labelNames' => ['foo', 'bar'],
+                            'samples' => [
+                                [
+                                    'labelValues' => ['lalal', 'lululu'],
                                     'value' => 3,
                                     'name' => 'test_some_metric',
-                                    'labelNames' => array()
-                                ),
-                            )
-                        )
-                    )
-                )
+                                    'labelNames' => [],
+                                ],
+                            ],
+                        ]
+                    ),
+                ]
             )
         );
     }
@@ -63,29 +65,29 @@ abstract class AbstractCounterTest extends PHPUnit_Framework_TestCase
      */
     public function itShouldIncreaseWithoutLabelWhenNoLabelsAreDefined()
     {
-        $gauge = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing');
-        $gauge->inc();
+        $counter = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing');
+        $counter->inc();
         $this->assertThat(
             $this->adapter->collect(),
             $this->equalTo(
-                array(
+                [
                     new MetricFamilySamples(
-                        array(
+                        [
                             'type' => Counter::TYPE,
                             'help' => 'this is for testing',
                             'name' => 'test_some_metric',
-                            'labelNames' => array(),
-                            'samples' => array(
-                                array(
-                                    'labelValues' => array(),
+                            'labelNames' => [],
+                            'samples' => [
+                                [
+                                    'labelValues' => [],
                                     'value' => 1,
                                     'name' => 'test_some_metric',
-                                    'labelNames' => array()
-                                ),
-                            )
-                        )
-                    )
-                )
+                                    'labelNames' => [],
+                                ],
+                            ],
+                        ]
+                    ),
+                ]
             )
         );
     }
@@ -95,50 +97,50 @@ abstract class AbstractCounterTest extends PHPUnit_Framework_TestCase
      */
     public function itShouldIncreaseTheCounterByAnArbitraryInteger()
     {
-        $gauge = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing', array('foo', 'bar'));
-        $gauge->inc(array('lalal', 'lululu'));
-        $gauge->incBy(123, array('lalal', 'lululu'));
+        $counter = new Counter($this->adapter, 'test', 'some_metric', 'this is for testing', ['foo', 'bar']);
+        $counter->inc(['lalal', 'lululu']);
+        $counter->incBy(123, ['lalal', 'lululu']);
         $this->assertThat(
             $this->adapter->collect(),
             $this->equalTo(
-                array(
+                [
                     new MetricFamilySamples(
-                        array(
+                        [
                             'type' => Counter::TYPE,
                             'help' => 'this is for testing',
                             'name' => 'test_some_metric',
-                            'labelNames' => array('foo', 'bar'),
-                            'samples' => array(
-                                array(
-                                    'labelValues' => array('lalal', 'lululu'),
+                            'labelNames' => ['foo', 'bar'],
+                            'samples' => [
+                                [
+                                    'labelValues' => ['lalal', 'lululu'],
                                     'value' => 124,
                                     'name' => 'test_some_metric',
-                                    'labelNames' => array()
-                                ),
-                            )
-                        )
-                    )
-                )
+                                    'labelNames' => [],
+                                ],
+                            ],
+                        ]
+                    ),
+                ]
             )
         );
     }
 
     /**
      * @test
-     * @expectedException \InvalidArgumentException
      */
     public function itShouldRejectInvalidMetricsNames()
     {
+        $this->expectException(InvalidArgumentException::class);
         new Counter($this->adapter, 'test', 'some metric invalid metric', 'help');
     }
 
     /**
      * @test
-     * @expectedException \InvalidArgumentException
      */
     public function itShouldRejectInvalidLabelNames()
     {
-        new Counter($this->adapter, 'test', 'some_metric', 'help', array('invalid label'));
+        $this->expectException(InvalidArgumentException::class);
+        new Counter($this->adapter, 'test', 'some_metric', 'help', ['invalid label']);
     }
 
     /**
@@ -150,30 +152,30 @@ abstract class AbstractCounterTest extends PHPUnit_Framework_TestCase
     public function isShouldAcceptAnySequenceOfBasicLatinCharactersForLabelValues($value)
     {
         $label = 'foo';
-        $histogram = new Counter($this->adapter, 'test', 'some_metric', 'help', array($label));
-        $histogram->inc(array($value));
+        $histogram = new Counter($this->adapter, 'test', 'some_metric', 'help', [$label]);
+        $histogram->inc([$value]);
 
         $metrics = $this->adapter->collect();
-        self::assertInternalType('array', $metrics);
-        self::assertCount(1, $metrics);
-        self::assertContainsOnlyInstancesOf(MetricFamilySamples::class, $metrics);
+        $this->assertIsArray($metrics);
+        $this->assertCount(1, $metrics);
+        $this->assertContainsOnlyInstancesOf(MetricFamilySamples::class, $metrics);
 
         $metric = reset($metrics);
         $samples = $metric->getSamples();
-        self::assertContainsOnlyInstancesOf(Sample::class, $samples);
+        $this->assertContainsOnlyInstancesOf(Sample::class, $samples);
 
         foreach ($samples as $sample) {
             $labels = array_combine(
                 array_merge($metric->getLabelNames(), $sample->getLabelNames()),
                 $sample->getLabelValues()
             );
-            self::assertEquals($value, $labels[$label]);
+            $this->assertEquals($value, $labels[$label]);
         }
     }
 
     /**
-     * @see isShouldAcceptArbitraryLabelValues
      * @return array
+     * @see isShouldAcceptArbitraryLabelValues
      */
     public function labelValuesDataProvider()
     {
@@ -181,10 +183,8 @@ abstract class AbstractCounterTest extends PHPUnit_Framework_TestCase
         // Basic Latin
         // See https://en.wikipedia.org/wiki/List_of_Unicode_characters#Basic_Latin
         for ($i = 32; $i <= 121; $i++) {
-            $cases['ASCII code ' . $i] = array(chr($i));
+            $cases['ASCII code ' . $i] = [chr($i)];
         }
         return $cases;
     }
-
-    public abstract function configureAdapter();
 }
