@@ -1,13 +1,13 @@
 <?php
 
-
 namespace Prometheus\Exception;
 
+use Exception;
 
 /**
  * Exception thrown if an error occurs during metrics registration.
  */
-class MetricsRegistrationException extends \Exception
+class MetricsRegistrationException extends Exception
 {
 
 }
