@@ -1,13 +1,13 @@
 <?php
 
-
 namespace Prometheus\Exception;
 
+use Exception;
 
 /**
  * Exception thrown if an error occurs during metrics storage.
  */
-class StorageException extends \Exception
+class StorageException extends Exception
 {
 
 }
