@@ -1,7 +1,8 @@
 <?php
 
-namespace Prometheus\Storage;
+declare(strict_types=1);
 
+namespace Prometheus\Storage;
 
 use Prometheus\MetricFamilySamples;
 use RuntimeException;
@@ -16,7 +17,7 @@ class InMemory implements Adapter
     /**
      * @return MetricFamilySamples[]
      */
-    public function collect()
+    public function collect(): array
     {
         $metrics = $this->internalCollect($this->counters);
         $metrics = array_merge($metrics, $this->internalCollect($this->gauges));
@@ -24,14 +25,17 @@ class InMemory implements Adapter
         return $metrics;
     }
 
-    public function flushMemory()
+    public function flushMemory(): void
     {
         $this->counters = [];
         $this->gauges = [];
         $this->histograms = [];
     }
 
-    private function collectHistograms()
+    /**
+     * @return array
+     */
+    private function collectHistograms(): array
     {
         $histograms = [];
         foreach ($this->histograms as $histogram) {
@@ -41,7 +45,7 @@ class InMemory implements Adapter
                 'help' => $metaData['help'],
                 'type' => $metaData['type'],
                 'labelNames' => $metaData['labelNames'],
-                'buckets' => $metaData['buckets']
+                'buckets' => $metaData['buckets'],
             ];
 
             // Add the Inf bucket so we can compute it later on
@@ -69,7 +73,7 @@ class InMemory implements Adapter
                             'name' => $metaData['name'] . '_bucket',
                             'labelNames' => ['le'],
                             'labelValues' => array_merge($decodedLabelValues, [$bucket]),
-                            'value' => $acc
+                            'value' => $acc,
                         ];
                     } else {
                         $acc += $histogramBuckets[$labelValues][$bucket];
@@ -77,7 +81,7 @@ class InMemory implements Adapter
                             'name' => $metaData['name'] . '_' . 'bucket',
                             'labelNames' => ['le'],
                             'labelValues' => array_merge($decodedLabelValues, [$bucket]),
-                            'value' => $acc
+                            'value' => $acc,
                         ];
                     }
                 }
@@ -87,7 +91,7 @@ class InMemory implements Adapter
                     'name' => $metaData['name'] . '_count',
                     'labelNames' => [],
                     'labelValues' => $decodedLabelValues,
-                    'value' => $acc
+                    'value' => $acc,
                 ];
 
                 // Add the sum
@@ -95,16 +99,19 @@ class InMemory implements Adapter
                     'name' => $metaData['name'] . '_sum',
                     'labelNames' => [],
                     'labelValues' => $decodedLabelValues,
-                    'value' => $histogramBuckets[$labelValues]['sum']
+                    'value' => $histogramBuckets[$labelValues]['sum'],
                 ];
-
             }
             $histograms[] = new MetricFamilySamples($data);
         }
         return $histograms;
     }
 
-    private function internalCollect(array $metrics)
+    /**
+     * @param array $metrics
+     * @return array
+     */
+    private function internalCollect(array $metrics): array
     {
         $result = [];
         foreach ($metrics as $metric) {
@@ -122,7 +129,7 @@ class InMemory implements Adapter
                     'name' => $metaData['name'],
                     'labelNames' => [],
                     'labelValues' => $this->decodeLabelValues($labelValues),
-                    'value' => $value
+                    'value' => $value,
                 ];
             }
             $this->sortSamples($data['samples']);
@@ -131,14 +138,18 @@ class InMemory implements Adapter
         return $result;
     }
 
-    public function updateHistogram(array $data)
+    /**
+     * @param array $data
+     * @return void
+     */
+    public function updateHistogram(array $data): void
     {
         // Initialize the sum
         $metaKey = $this->metaKey($data);
         if (array_key_exists($metaKey, $this->histograms) === false) {
             $this->histograms[$metaKey] = [
                 'meta' => $this->metaData($data),
-                'samples' => []
+                'samples' => [],
             ];
         }
         $sumKey = $this->histogramBucketValueKey($data, 'sum');
@@ -164,14 +175,17 @@ class InMemory implements Adapter
         $this->histograms[$metaKey]['samples'][$bucketKey] += 1;
     }
 
-    public function updateGauge(array $data)
+    /**
+     * @param array $data
+     */
+    public function updateGauge(array $data): void
     {
         $metaKey = $this->metaKey($data);
         $valueKey = $this->valueKey($data);
         if (array_key_exists($metaKey, $this->gauges) === false) {
             $this->gauges[$metaKey] = [
                 'meta' => $this->metaData($data),
-                'samples' => []
+                'samples' => [],
             ];
         }
         if (array_key_exists($valueKey, $this->gauges[$metaKey]['samples']) === false) {
@@ -184,14 +198,17 @@ class InMemory implements Adapter
         }
     }
 
-    public function updateCounter(array $data)
+    /**
+     * @param array $data
+     */
+    public function updateCounter(array $data): void
     {
         $metaKey = $this->metaKey($data);
         $valueKey = $this->valueKey($data);
         if (array_key_exists($metaKey, $this->counters) === false) {
             $this->counters[$metaKey] = [
                 'meta' => $this->metaData($data),
-                'samples' => []
+                'samples' => [],
             ];
         }
         if (array_key_exists($valueKey, $this->counters[$metaKey]['samples']) === false) {
@@ -206,18 +223,16 @@ class InMemory implements Adapter
 
     /**
      * @param array $data
-     *
-     * @param       $bucket
-     *
+     * @param string $bucket
      * @return string
      */
-    private function histogramBucketValueKey(array $data, $bucket)
+    private function histogramBucketValueKey(array $data, $bucket): string
     {
         return implode(':', [
             $data['type'],
             $data['name'],
             $this->encodeLabelValues($data['labelValues']),
-            $bucket
+            $bucket,
         ]);
     }
 
@@ -226,9 +241,13 @@ class InMemory implements Adapter
      *
      * @return string
      */
-    private function metaKey(array $data)
+    private function metaKey(array $data): string
     {
-        return implode(':', [$data['type'], $data['name'], 'meta']);
+        return implode(':', [
+            $data['type'],
+            $data['name'],
+            'meta'
+        ]);
     }
 
     /**
@@ -236,10 +255,14 @@ class InMemory implements Adapter
      *
      * @return string
      */
-    private function valueKey(array $data)
+    private function valueKey(array $data): string
     {
-        return implode(':',
-            [$data['type'], $data['name'], $this->encodeLabelValues($data['labelValues']), 'value']);
+        return implode(':', [
+            $data['type'],
+            $data['name'],
+            $this->encodeLabelValues($data['labelValues']),
+            'value'
+        ]);
     }
 
     /**
@@ -247,7 +270,7 @@ class InMemory implements Adapter
      *
      * @return array
      */
-    private function metaData(array $data)
+    private function metaData(array $data): array
     {
         $metricsMetaData = $data;
         unset($metricsMetaData['value']);
@@ -256,7 +279,10 @@ class InMemory implements Adapter
         return $metricsMetaData;
     }
 
-    private function sortSamples(array &$samples)
+    /**
+     * @param array $samples
+     */
+    private function sortSamples(array &$samples): void
     {
         usort($samples, function ($a, $b) {
             return strcmp(implode("", $a['labelValues']), implode("", $b['labelValues']));
@@ -268,7 +294,7 @@ class InMemory implements Adapter
      * @return string
      * @throws RuntimeException
      */
-    private function encodeLabelValues(array $values)
+    private function encodeLabelValues(array $values): string
     {
         $json = json_encode($values);
         if (false === $json) {
@@ -282,7 +308,7 @@ class InMemory implements Adapter
      * @return array
      * @throws RuntimeException
      */
-    private function decodeLabelValues($values)
+    private function decodeLabelValues($values): array
     {
         $json = base64_decode($values, true);
         if (false === $json) {
