@@ -1,9 +1,7 @@
 <?php
 
-
 namespace Test\Prometheus\Redis;
 
-use function class_exists;
 use Prometheus\Storage\Redis;
 use Test\Prometheus\AbstractCollectorRegistryTest;
 
@@ -14,7 +12,7 @@ class CollectorRegistryTest extends AbstractCollectorRegistryTest
 {
     public function configureAdapter()
     {
-        $this->adapter = new Redis(array('host' => REDIS_HOST));
+        $this->adapter = new Redis(['host' => REDIS_HOST]);
         $this->adapter->flushRedis();
     }
 }
