@@ -9,11 +9,16 @@ use Test\Prometheus\AbstractHistogramTest;
  * See https://prometheus.io/docs/instrumenting/exposition_formats/
  * @requires extension redis
  */
-class HistogramTest extends AbstractHistogramTest
+class HistogramWithPrefixTest extends AbstractHistogramTest
 {
     public function configureAdapter()
     {
-        $this->adapter = new Redis(['host' => REDIS_HOST]);
+        $connection = new \Redis();
+        $connection->connect(REDIS_HOST);
+
+        $connection->setOption(\Redis::OPT_PREFIX, 'prefix:');
+
+        $this->adapter = Redis::fromExistingConnection($connection);
         $this->adapter->flushRedis();
     }
 }
