@@ -1,22 +1,39 @@
 <?php
 
-
 namespace Prometheus\Storage;
+
+use PHPUnit\Framework\TestCase;
+use Prometheus\Exception\StorageException;
 
 /**
  * @requires extension redis
  */
-class RedisTest extends \PHPUnit_Framework_TestCase
+class RedisTest extends TestCase
 {
     /**
      * @test
-     * @expectedException \Prometheus\Exception\StorageException
-     * @expectedExceptionMessage Can't connect to Redis server
      */
     public function itShouldThrowAnExceptionOnConnectionFailure()
     {
-        $redis = new Redis(array('host' => 'doesntexist.test'));
+        $redis = new Redis(['host' => '/dev/null']);
+
+        $this->expectException(StorageException::class);
+        $this->expectExceptionMessage("Can't connect to Redis server");
+
+        $redis->collect();
         $redis->flushRedis();
     }
 
+    /**
+     * @test
+     */
+    public function itShouldThrowExceptionWhenInjectedRedisIsNotConnected()
+    {
+        $connection = new \Redis();
+
+        $this->expectException(StorageException::class);
+        $this->expectExceptionMessage('Connection to Redis server not established');
+
+        Redis::fromExistingConnection($connection);
+    }
 }
